% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{fit_inlami}
\alias{fit_inlami}
\title{Fit model for measurement error and missing data in INLA}
\usage{
fit_inlami(
  formula_moi,
  formula_imp,
  family_moi,
  data,
  error_type = "classical",
  classical_error_scaling = NULL,
  prior.prec.y = NULL,
  prior.prec.u_b = NULL,
  prior.prec.u_c = NULL,
  prior.prec.r = NULL,
  initial.prec.y = NULL,
  initial.prec.u_b = NULL,
  initial.prec.u_c = NULL,
  initial.prec.r = NULL,
  control.family = NULL,
  control.predictor = NULL,
  ...
)
}
\arguments{
\item{formula_moi}{an object of class "formula", describing the main model to be fitted.}

\item{formula_imp}{an object of class "formula", describing the imputation model for the mismeasured and/or missing observations.}

\item{family_moi}{a string indicating the likelihood family for the model of interest (the main model).}

\item{data}{a data frame containing the variables in the model.}

\item{error_type}{type of error (one or more of "classical", "berkson", "missing")}

\item{classical_error_scaling}{can be specified if the classical measurement error varies across observations. Must be a vector of the same length as the data.}

\item{prior.prec.y}{a string containing the parameters for the prior for the precision of the residual term for the model of interest.}

\item{prior.prec.u_b}{a string containing the parameters for the prior for the precision of the error term for the Berkson error model.}

\item{prior.prec.u_c}{a string containing the parameters for the prior for the precision of the error term for the classical error model.}

\item{prior.prec.r}{a string containing the parameters for the precision of the latent variable r, which is the variable being described in the imputation model.}

\item{initial.prec.y}{the initial value for the precision of the residual term for the model of interest.}

\item{initial.prec.u_b}{the initial value for the precision of the residual term for the Berkson error term.}

\item{initial.prec.u_c}{the initial value for the precision of the residual term for the classical error term.}

\item{initial.prec.r}{the initial value for the precision of the residual term for the latent variable r.}

\item{control.family}{control.family for use in inla (can be provided directly instead of passing the "prior.prec...." and "initial.prec..." arguments.}

\item{control.predictor}{control.predictor for use in inla.}

\item{...}{other arguments to pass to `inla`.}
}
\value{
An object of class \code{inlami}.
}
\description{
A wrapper function around "INLA::inla()", providing the necessary structure to fit the hierarchical measurement error model that adjusts coefficient estimates to account for biases due to measurement error and missing data.
}
\examples{
simple_moi <- y ~ x + z
simple_imp <- x ~ z

# Prior for beta.x
prior.beta <- c(0, 1/1000) # N(0, 10^3)

# Priors for y, measurement error and true x-value precision
prior.prec.y <- c(0.5, 0.5) # Gamma(0.5, 0.5)
prior.prec.u_b <- c(10, 9) # Gamma(0.5, 0.5)
prior.prec.u_c <- c(10, 9) # Gamma(0.5, 0.5)
prior.prec.r <- c(0.5, 0.5) # Gamma(0.5, 0.5)

# Initial values
initial.prec.y <- 1
initial.prec.u_b <- 1
initial.prec.u_c <- 1
initial.prec.r <- 1

# Fit the model
simple_model <- fit_inlami(data = simple_data,
                         formula_moi = simple_moi,
                         formula_imp = simple_imp,
                         family_moi = "gaussian",
                         error_type = c("berkson", "classical"),
                         prior.prec.y = prior.prec.y,
                         prior.prec.u_b = prior.prec.u_b,
                         prior.prec.u_c = prior.prec.u_c,
                         prior.prec.r = prior.prec.r,
                         initial.prec.y = initial.prec.y,
                         initial.prec.u_b = initial.prec.u_b,
                         initial.prec.u_c = initial.prec.u_c,
                         initial.prec.r = initial.prec.r)
}
