% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{inla_survival_families}
\alias{inla_survival_families}
\title{List the survival likelihoods in INLA}
\usage{
inla_survival_families()
}
\value{
List of survival models in INLA
}
\description{
List the survival likelihoods in INLA
}
