% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{make_inlami_formula}
\alias{make_inlami_formula}
\title{Make formula for measurement error and missing data model}
\usage{
make_inlami_formula(formula_moi, formula_imp, error_type = "classical")
}
\arguments{
\item{formula_moi}{an object of class "formula", describing the main model to be fitted.}

\item{formula_imp}{an object of class "formula", describing the imputation model for the mismeasured and/or missing observations.}

\item{error_type}{Type of error (one of "classical", "berkson", "missing")}
}
\value{
An object of class "formula".
}
\description{
Make formula for measurement error and missing data model
}
\examples{
f_moi <- y ~ x + z
f_imp <- x ~ z
make_inlami_formula(formula_moi = f_moi, formula_imp = f_imp, error_type = "classical")
}
