% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{make_inlami_matrices}
\alias{make_inlami_matrices}
\title{Make matrices for joint model specification in INLA}
\usage{
make_inlami_matrices(data, formula_moi, formula_imp, error_type = "classical")
}
\arguments{
\item{data}{A data frame with response and covariate variables for the main model and the imputation model.}

\item{formula_moi}{an object of class "formula", describing the main model to be fitted.}

\item{formula_imp}{an object of class "formula", describing the imputation model for the mismeasured and/or missing observations.}

\item{error_type}{Type of error (one or more of "classical", "berkson", "missing")}
}
\value{
A data frame with all the data needed for the model, structured in the right way for INLA.
}
\description{
Make matrices for joint model specification in INLA
}
