% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{make_inlami_scaling_vector}
\alias{make_inlami_scaling_vector}
\title{Construct scaling vector to scale the precision of non-mismeasured observations}
\usage{
make_inlami_scaling_vector(
  inlami_stack,
  error_type,
  classical_error_scaling = NULL
)
}
\arguments{
\item{inlami_stack}{an object of class inlami.data.stack containing data structured}

\item{error_type}{one of "classical", "berkson" or "missing".}

\item{classical_error_scaling}{can be specified if the classical measurement error varies across observations. Must be a vector of the same length as the data.}
}
\value{
A vector reflecting the scaling factor for the residual terms in each model level.
}
\description{
Construct scaling vector to scale the precision of non-mismeasured observations
}
\examples{
stacks <- make_inlami_stacks(data = simple_data,
                             formula_moi = y ~ x + z,
                             formula_imp = x ~ z,
                             error_type = c("classical", "berkson"))
make_inlami_scaling_vector(stacks,
                           error_type = c("classical", "berkson"))
}
