% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhanes_survival}
\alias{nhanes_survival}
\title{Survival data with repeated systolic blood pressure measurements}
\format{
## `nhanes_survival`
A data frame with 3433 rows and 8 columns:
\describe{
  \item{sbp1}{systolic blood pressure (standardized), first measurement}
  \item{sbp2}{systolic blood pressure (standardized), second measurement}
  \item{sex}{sex (0 = female, 1 = male)}
  \item{age}{age (standardized)}
  \item{smoke}{smoking status (0 = no, 1 = yes)}
  \item{diabetes}{diabetes status (0 = no, 1 = yes)}
  \item{d}{censoring status (0 = censored, 1 = observed death due to cardiovascular disease)}
  \item{t}{time until death due to cardiovascular disease occurs}
}
}
\source{
https://github.com/ruthkeogh/meas_error_handbook
}
\usage{
nhanes_survival
}
\description{
A dataset containing a repeated blood pressure measurement along with some other variables for participants in the Third National Health and Nutrition Survey (NHANES III), merged with data from the US National Death Index by Ruth H. Keogh and Jonathan Bartlett. For the illustration purposes in this package, we have left out observations where smoking status is missing.
}
\keyword{datasets}
