% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_inlami}
\alias{plot_inlami}
\title{Plot model summary}
\usage{
plot_inlami(
  inlami_model,
  plot_moi = TRUE,
  plot_imp = TRUE,
  plot_intercepts = TRUE
)
}
\arguments{
\item{inlami_model}{the model returned from the fit_inlami function.}

\item{plot_moi}{should the posterior mean for the coefficients of the model of interest be plotted? Defaults to TRUE.}

\item{plot_imp}{should the posterior mean for the coefficients of the imputation model be plotted? Defaults to TRUE.}

\item{plot_intercepts}{should the posterior mean for the intercept(s) be plotted?}
}
\value{
An object of class "ggplot2" that plots the posterior mean and 95 \% credible interval for each coefficient in the model. The coefficients are colored to indicate if they belong to the main or imputation model, and the variable with error is also highlighted.
}
\description{
Plot model summary
}
\examples{
#' #' simple_moi <- y ~ x + z
simple_imp <- x ~ z

# Prior for beta.x
prior.beta <- c(0, 1/1000) # N(0, 10^3)

# Priors for y, measurement error and true x-value precision
prior.prec.y <- c(0.5, 0.5) # Gamma(0.5, 0.5)
prior.prec.u_b <- c(10, 9) # Gamma(0.5, 0.5)
prior.prec.u_c <- c(10, 9) # Gamma(0.5, 0.5)
prior.prec.r <- c(0.5, 0.5) # Gamma(0.5, 0.5)

# Initial values
initial.prec.y <- 1
initial.prec.u_b <- 1
initial.prec.u_c <- 1
initial.prec.r <- 1

# Fit the model
#simple_model <- fit_inlami(data = simple_data,
#                         formula_moi = simple_moi,
#                         formula_imp = simple_imp,
#                         family_moi = "gaussian",
#                         error_type = c("berkson", "classical"),
#                         prior.prec.y = prior.prec.y,
#                         prior.prec.u_b = prior.prec.u_b,
#                         prior.prec.u_c = prior.prec.u_c,
#                         prior.prec.r = prior.prec.r,
#                         initial.prec.y = initial.prec.y,
#                         initial.prec.u_b = initial.prec.u_b,
#                         initial.prec.u_c = initial.prec.u_c,
#                         initial.prec.r = initial.prec.r)

#plot_inlami(simple_model)
}
