% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{show_data_structure}
\alias{show_data_structure}
\title{Visualize the model data structure as matrices}
\usage{
show_data_structure(stack)
}
\arguments{
\item{stack}{}
}
\value{
A latex table?
}
\description{
Visualize the model data structure as matrices
}
