% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_data}
\alias{simple_data}
\title{Simple simulated data}
\format{
## `simple_data`
A data frame with 1000 rows and 4 columns:
\describe{
  \item{y}{Response variable}
  \item{x}{Covariate measured with error, both Berkson and classical error and missing observations}
  \item{x_true}{Correct version of the covariate with error}
  \item{z}{Error free covariate, correlated with x}
}
}
\source{
The dataset is simulated.
}
\usage{
simple_data
}
\description{
A simulated dataset to demonstrate how to model different types of measurement error and missing data using the 'inlami' package.
}
\keyword{datasets}
