% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summarise_inlami}
\alias{summarise_inlami}
\title{Summarise INLA measurement error and missing data model}
\usage{
summarise_inlami(inlami_model)
}
\arguments{
\item{inlami_model}{the model returned from the fit_inlami function.}
}
\value{
Prints a structured summary tailored to clarify the different components of the measurement error/missing data model.
}
\description{
Summarise INLA measurement error and missing data model
}
\examples{
simple_moi <- y ~ x + z
simple_imp <- x ~ z

# Prior for beta.x
prior.beta <- c(0, 1/1000) # N(0, 10^3)

# Priors for y, measurement error and true x-value precision
prior.prec.y <- c(0.5, 0.5) # Gamma(0.5, 0.5)
prior.prec.u_b <- c(10, 9) # Gamma(0.5, 0.5)
prior.prec.u_c <- c(10, 9) # Gamma(0.5, 0.5)
prior.prec.r <- c(0.5, 0.5) # Gamma(0.5, 0.5)

# Initial values
initial.prec.y <- 1
initial.prec.u_b <- 1
initial.prec.u_c <- 1
initial.prec.r <- 1

# Fit the model
#simple_model <- fit_inlami(data = simple_data,
#                         formula_moi = simple_moi,
#                         formula_imp = simple_imp,
#                         family_moi = "gaussian",
#                         error_type = c("berkson", "classical"),
#                         prior.prec.y = prior.prec.y,
#                         prior.prec.u_b = prior.prec.u_b,
#                         prior.prec.u_c = prior.prec.u_c,
#                         prior.prec.r = prior.prec.r,
#                         initial.prec.y = initial.prec.y,
#                         initial.prec.u_b = initial.prec.u_b,
#                         initial.prec.u_c = initial.prec.u_c,
#                         initial.prec.r = initial.prec.r)

#summarise_inlami(simple_model)
}
