% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{make_inlami_scaling_vector}
\alias{make_inlami_scaling_vector}
\title{Construct scaling vector to scale the precision of non-mismeasured observations}
\usage{
make_inlami_scaling_vector(data, error_variable, error_type)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}

\item{error_variable}{the variable with measurement error or missingness.}

\item{error_type}{one of "classical", "berkson" or "missing".}
}
\value{
A vector reflecting the scaling factor for the residual terms in each model level.
}
\description{
Construct scaling vector to scale the precision of non-mismeasured observations
}
\examples{
make_inlami_scaling_vector(simple_data,
                           error_variable = "x",
                           error_type = c("classical", "berkson"))
}
