% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{make_inlami_stacks}
\alias{make_inlami_stacks}
\title{Make data stacks for joint model specification in INLA}
\usage{
make_inlami_stacks(
  data,
  formula_moi,
  formula_imp,
  error_type = "classical",
  repeated_observations = FALSE
)
}
\arguments{
\item{data}{A data frame with response and covariate variables for the main model and the imputation model.}

\item{formula_moi}{an object of class "formula", describing the main model to be fitted.}

\item{formula_imp}{an object of class "formula", describing the imputation model for the mismeasured and/or missing observations.}

\item{error_type}{Type of error (one or more of "classical", "berkson", "missing")}

\item{repeated_observations}{Does the variable with measurement error and/or missingness have repeated observations? If so, set this to "TRUE". In that case, when specifying the formula, use the name of the variable without any numbers, but when specifying the data, make sure that the repeated measurements end in a number, i.e "sbp1" and "sbp2".}
}
\value{
An object of class inla.stack with data structured according to specified formulas and error models.
}
\description{
Make data stacks for joint model specification in INLA
}
\examples{
f_moi <- y ~ x + z
f_imp <- x ~ z
make_inlami_stacks(data = simple_data,
              formula_moi = f_moi,
              formula_imp = f_imp,
              error_type = "classical")
}
