% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary.inlami}
\alias{summary.inlami}
\alias{print.summary.inlami}
\title{Summary method for inlami}
\usage{
\method{summary}{inlami}(object, ...)

\method{print}{summary.inlami}(x, ...)
}
\arguments{
\item{object}{model of class `inlami`.}

\item{...}{other arguments}

\item{x}{object of class summary.inlami.}
}
\value{
`summary.inlami` returns an object of class `summary.inlami`, a list of components to print.
}
\description{
Takes a fitted `inlami` object produced by
`fit_inlami` and produces a summary from it.
}
\examples{
simple_moi <- y ~ x + z
simple_imp <- x ~ z

# Fit the model
simple_model <- fit_inlami(data = simple_data,
                           formula_moi = simple_moi,
                           formula_imp = simple_imp,
                           family_moi = "gaussian",
                           error_type = c("berkson", "classical"),
                           prior.prec.moi = c(10, 9),
                           prior.prec.berkson = c(10, 9),
                           prior.prec.classical = c(10, 9),
                           prior.prec.imp = c(10, 9),
                           initial.prec.moi = 1,
                           initial.prec.berkson = 1,
                           initial.prec.classical = 1,
                           initial.prec.imp = 1)

summary(simple_model)
}
