% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary.inlami}
\alias{summary.inlami}
\alias{print.summary.inlami}
\title{Summary method for inlami}
\usage{
\method{summary}{inlami}(object, ...)

\method{print}{summary.inlami}(x, ...)
}
\arguments{
\item{object}{model of class `inlami`.}

\item{...}{other arguments}

\item{x}{object of class summary.inlami.}
}
\value{
`summary.inlami` returns an object of class `summary.inlami`, a list of components to print.
}
\description{
Takes a fitted `inlami` object produced by
`fit_inlami` and produces a summary from it.
}
\examples{
simple_moi <- y ~ x + z
simple_imp <- x ~ z

# Prior for beta.x
prior.beta <- c(0, 1/1000) # N(0, 10^3)

# Priors for y, measurement error and true x-value precision
prior.prec.y <- c(0.5, 0.5) # Gamma(0.5, 0.5)
prior.prec.u_b <- c(10, 9) # Gamma(0.5, 0.5)
prior.prec.u_c <- c(10, 9) # Gamma(0.5, 0.5)
prior.prec.r <- c(0.5, 0.5) # Gamma(0.5, 0.5)

# Initial values
initial.prec.y <- 1
initial.prec.u_b <- 1
initial.prec.u_c <- 1
initial.prec.r <- 1

# Fit the model
simple_model <- fit_inlami(data = simple_data,
                         formula_moi = simple_moi,
                         formula_imp = simple_imp,
                         family_moi = "gaussian",
                         error_type = c("berkson", "classical"),
                         prior.prec.y = prior.prec.y,
                         prior.prec.u_b = prior.prec.u_b,
                         prior.prec.u_c = prior.prec.u_c,
                         prior.prec.r = prior.prec.r,
                         initial.prec.y = initial.prec.y,
                         initial.prec.u_b = initial.prec.u_b,
                         initial.prec.u_c = initial.prec.u_c,
                         initial.prec.r = initial.prec.r)

summary(simple_model)
}
